pyproject.toml
[tool.poetry]
name = "medical_psychology_poeagent"
version = "0.1.0"
description = "Medical Psychology RAG Agent with Supervisor Pattern"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{ include = "medical_psychology_agent", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.11, <4.0.0"
langchain = ">=1.2.0,<2.0.0"
langchain-openai = ">=1.1.5,<2.0.0"
langchain-community = ">=0.3.0,<1.0.0"
langchain-qdrant = ">=1.1.0,<2.0.0"
langgraph = ">=1.0.2,<1.1.0"
langfuse = "3.10.1"
qdrant-client = ">=1.12.0,<2.0.0"
datasets = ">=3.0.0,<4.0.0"
streamlit = ">=1.52.2,<2.0.0"
python-dotenv = ">=1.2.1,<2.0.0"
tiktoken = ">=0.8.0,<1.0.0"
pandas = ">=2.3.3,<3.0.0"
cohere = ">=5.0.0,<6.0.0"
packaging = "<25"

[tool.poetry.group.dev.dependencies]
black = "^24.0.0"
flake8 = "^7.0.0"
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"